Project 구조 변경
- View의 실행 로직 변경
- Project 구조 변경
- BaseActivity, BaseSideEffect, BaseState, BaseViewModel 삭제
