+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:name=".EggtartApplication"
         tools:targetApi="31">
         <activity
-            android:name=".features.main.MainActivity"
+            android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.Eggtart">
