+++ app/src/main/java/com/eggtart/eggtart/MainSideEffect.kt
@@ -1,0 +1,6 @@
+package com.eggtart.eggtart
+
+/**
+ *  Created by wonjin on 2024/03/29
+ **/
+sealed class MainSideEffect {}
