+++ app/src/main/java/com/eggtart/eggtart/MainState.kt
@@ -1,0 +1,6 @@
+package com.eggtart.eggtart
+
+/**
+ *  Created by wonjin on 2024/03/29
+ **/
+class MainState
