+++ app/src/main/java/com/eggtart/eggtart/MainViewModel.kt
@@ -1,0 +1,14 @@
+package com.eggtart.eggtart
+
+import androidx.lifecycle.ViewModel
+import org.orbitmvi.orbit.ContainerHost
+import org.orbitmvi.orbit.viewmodel.container
+
+/**
+ *  Created by wonjin on 2024/03/21
+ **/
+
+class MainViewModel : ContainerHost<MainState, MainSideEffect>, ViewModel() {
+    override val container = container<MainState, MainSideEffect>(MainState())
+
+}
