+++ app/src/main/java/com/eggtart/eggtart/features/home/mandalart/MandalartSideEffect.kt
@@ -1,0 +1,7 @@
+package com.eggtart.eggtart.features.home.mandalart
+
+/**
+ *  Created by wonjin on 2024/03/29
+ **/
+sealed class MandalartSideEffect {
+}
