--- app/src/main/java/com/eggtart/eggtart/features/main/screens/timer/TimerScreen.kt
@@ -1,19 +1,0 @@
-package com.eggtart.eggtart.features.main.screens.timer
-
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-
-/**
- * Created by 노원진 on 2024.03.24
- */
-
-@Composable
-fun TimerScreen() {
-    Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
-        Text(text = "Timer")
-    }
-}
